+++ core/lib/Thelia/Controller/Admin/ProductController.php
@@ -1061,7 +1061,16 @@ class ProductController extends AbstractSeoCrudController
 
         // Log object modification
         if (null !== $changedObject = $event->getProductSaleElement()) {
-            $this->adminLogAppend($this->resourceCode, AccessManager::UPDATE, sprintf("Product Sale Element (ID %s) for product reference %s modified", $changedObject->getId(), $event->getProduct()->getRef()));
+            $this->adminLogAppend(
+                $this->resourceCode,
+                AccessManager::UPDATE,
+                sprintf(
+                    "Product Sale Element (ID %s) for product reference %s modified",
+                    $changedObject->getId(),
+                    $event->getProduct()->getRef()
+                ),
+                $changedObject->getId()
+            );
         }
     }
 
@@ -1237,7 +1246,15 @@ class ProductController extends AbstractSeoCrudController
             $this->dispatch(TheliaEvents::PRODUCT_COMBINATION_GENERATION, $event);
 
             // Log object modification
-            $this->adminLogAppend($this->resourceCode, AccessManager::CREATE, sprintf("Combination generation for product reference %s", $event->getProduct()->getRef()));
+            $this->adminLogAppend(
+                $this->resourceCode,
+                AccessManager::CREATE,
+                sprintf(
+                    "Combination generation for product reference %s",
+                    $event->getProduct()->getRef()
+                ),
+                $event->getProduct()->getId()
+            );
 
            // Redirect to the success URL
             return $this->generateSuccessRedirect($changeForm);