+++ core/lib/Thelia/Core/Form/Type/Field/FormFirewallIdType.php
@@ -1,0 +1,43 @@
+<?php
+/*************************************************************************************/
+/*      This file is part of the Thelia package.                                     */
+/*                                                                                   */
+/*      Copyright (c) OpenStudio                                                     */
+/*      email : dev@thelia.net                                                       */
+/*      web : http://www.thelia.net                                                  */
+/*                                                                                   */
+/*      For the full copyright and license information, please view the LICENSE.txt  */
+/*      file that was distributed with this source code.                             */
+/*************************************************************************************/
+
+namespace Thelia\Core\Form\Type\Field;
+
+use Thelia\Model\FormFirewallQuery;
+
+/**
+ * Class FormFirewallIdType
+ * @package Thelia\Core\Form\Type\Field
+ * @author Julien Chanséaume <jchanseaume@openstudio.fr>
+ */
+class FormFirewallIdType extends AbstractIdType
+{
+    /**
+     * @return \Propel\Runtime\ActiveQuery\ModelCriteria
+     *
+     * Get the model query to check
+     */
+    protected function getQuery()
+    {
+        return new FormFirewallQuery();
+    }
+
+    /**
+     * Returns the name of this type.
+     *
+     * @return string The name of this type
+     */
+    public function getName()
+    {
+        return "form_firewall_id";
+    }
+}